Allow for macros that are defined in modules
Currently, it is not possible to do
```julia
@mix Stipple.@with_kw mutable struct reactive
  Stipple.@reactors
end
```
as `Stipple.@with_kw` is not a Symbol but an expression. Therefore, it is currently not possible to use macros for mixing that are not available in the current name space. Changing the array type as proposed cures that short coming.